update in add toy and update toy page

+++ src/pages/AddToys/AddToys.jsx
@@ -15,7 +15,7 @@ const AddToys = () => {
     formState: { errors },
   } = useForm();
   const onSubmit = (data) => {
-    data.price = parseInt(data.price)
+    data.price = parseFloat(data.price)
     console.log(data);
     fetch(`https://assignment-eleven-server-alpha.vercel.app/toys`, {
       method: "POST",
